+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Update dependency pytest-asyncio to v1
+
 ## [0.14.1] - 2025-05-31
 
 - handle missing MQTT addon [#438](https://github.com/s-allius/tsun-gen3-proxy/issues/438)
-
-## [0.14.0] - 2025-05-29
 
 ## [0.14.0] - 2025-05-29