+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Update dependency pytest-asyncio to v1
+
+## [0.14.1] - 2025-05-31
+
+- handle missing MQTT addon [#438](https://github.com/s-allius/tsun-gen3-proxy/issues/438)
+
+## [0.14.0] - 2025-05-29
+
+- add-on: bump python to version 3.12.10-r1
+- set no of pv modules for MS800 GEN3PLUS inverters
+- fix the paths to copy the config.example.toml file during proxy start
+- add MQTT topic `dcu_power` for setting output power on DCUs
 - add MQTT topic `dcu_power` for setting output power on DCUs
 - Update ghcr.io/hassio-addons/base Docker tag to v17.2.5
 - fix a lot of pytest-asyncio problems in the unit tests